twig_your_health/requirements.txt
aioredis==1.0.0
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.1.0
autobahn==18.5.1
Automat==0.6.0
Babel==2.5.3
beautifulsoup4==4.6.0
boto3==1.7.28
botocore==1.10.28
certifi==2018.1.18
cffi==1.11.5
channels==2.1.1
channels-redis==2.1.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
daphne==2.1.2
dj-database-url==0.5.0
Django==2.0.5
django-annoying==0.10.4
django-appconf==1.0.2
django-compressor==2.2
django-extensions==2.0.6
django-js-asset==1.0.0
django-material==1.2.2
django-mptt==0.9.0
django-phonenumber-field==2.0.0
django-sass-processor==0.5.8
django-storages==1.6.1
docutils==0.14
gunicorn==19.7.1
hiredis==0.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
jmespath==0.9.3
libsass==0.13.7
lxml==4.1.1
msgpack==0.5.6
parsel==1.4.0
phonenumberslite==8.9.0
Pillow==5.0.0
psycopg2==2.7.4
pushbullet.py==0.11.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.5.0
pyparsing==2.2.0
python-dateutil==2.7.3
python-magic==0.4.15
pytz==2018.4
queuelib==1.4.2
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
s3transfer==0.1.13
Scrapy==1.5.0
scrapyd==1.2.0
selenium==3.10.0
service-identity==17.0.0
six==1.11.0
SkPy==0.8.1
sorl-thumbnail==12.4.1
Twisted==18.4.0
txaio==2.10.0
urllib3==1.22
w3lib==1.19.0
websocket-client==0.47.0
whitenoise==3.3.1
zope.interface==4.5.0
